% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.ks-package}
\alias{mlr3learners.ks}
\alias{mlr3learners.ks-package}
\title{mlr3learners.ks: Learners from the {ks} package for 'mlr3'}
\description{
Adds learners from the {ks} package to 'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
